Refactor code to separate interace and models to api module and implementaiton of api to impl module
