+++ order-management-service-api/src/main/java/com/yrlalal/ordermanagementservice/api/v1/OrderManagementServiceApiConfiguration.java
@@ -1,0 +1,7 @@
+package com.yrlalal.ordermanagementservice.api.v1;
+
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class OrderManagementServiceApiConfiguration {
+}
