+++ order-management-service-api/src/main/java/com/yrlalal/ordermanagementservice/api/v1/model/OrderItemPK.java
@@ -1,0 +1,14 @@
+package com.yrlalal.ordermanagementservice.api.v1.model;
+
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+
+import java.io.Serializable;
+
+public class OrderItemPK implements Serializable {
+    private Integer orderItemId;
+
+    @ManyToOne
+    @JoinColumn(name = "order_id")
+    private Order order;
+}
