+++ order-management-service-api/src/main/java/com/yrlalal/ordermanagementservice/api/v1/model/OrderStatus.java
@@ -1,0 +1,5 @@
+package com.yrlalal.ordermanagementservice.api.v1.model;
+
+public enum OrderStatus {
+    NEW, CHECKOUT, ACTIVE, COMPLETED, CANCELLATION_IN_PROGRESS, CANCELLED;
+}
