+++ order-management-service-impl/src/main/java/com/yrlalal/ordermanagementservice/impl/exception/BadRequestException.java
@@ -1,0 +1,7 @@
+package com.yrlalal.ordermanagementservice.impl.exception;
+
+public class BadRequestException extends RuntimeException {
+    public BadRequestException(String message) {
+        super(message);
+    }
+}
