+++ order-management-service-impl/src/main/java/com/yrlalal/ordermanagementservice/impl/exception/OrderNotFoundException.java
@@ -1,0 +1,7 @@
+package com.yrlalal.ordermanagementservice.impl.exception;
+
+public class OrderNotFoundException extends RuntimeException {
+    public OrderNotFoundException(String message) {
+        super(message);
+    }
+}
