+++ order-management-service-impl/src/main/java/com/yrlalal/ordermanagementservice/impl/v1/repository/OrderItemRepository.java
@@ -1,0 +1,8 @@
+package com.yrlalal.ordermanagementservice.impl.v1.repository;
+
+import com.yrlalal.ordermanagementservice.api.v1.model.OrderItem;
+import com.yrlalal.ordermanagementservice.api.v1.model.OrderItemPK;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OrderItemRepository extends JpaRepository<OrderItem, OrderItemPK> {
+}
