+++ order-management-service-impl/src/main/java/com/yrlalal/ordermanagementservice/impl/v1/repository/OrderRepository.java
@@ -1,0 +1,7 @@
+package com.yrlalal.ordermanagementservice.impl.v1.repository;
+
+import com.yrlalal.ordermanagementservice.api.v1.model.Order;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OrderRepository extends JpaRepository<Order, String> {
+}
