+++ order-management-service-impl/src/test/java/com/yrlalal/ordermanagementservice/impl/OrderManagementServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.yrlalal.ordermanagementservice.impl;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class OrderManagementServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
