--- src/main/java/com/yrlalal/ordermanagementservice/OrderManagementServiceApplication.java
@@ -1,13 +1,0 @@
-package com.yrlalal.ordermanagementservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class OrderManagementServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(OrderManagementServiceApplication.class, args);
-	}
-
-}
