--- src/main/java/com/yrlalal/ordermanagementservice/exception/BadRequestException.java
@@ -1,7 +1,0 @@
-package com.yrlalal.ordermanagementservice.exception;
-
-public class BadRequestException extends RuntimeException {
-    public BadRequestException(String message) {
-        super(message);
-    }
-}
