--- src/main/java/com/yrlalal/ordermanagementservice/exception/OrderNotFoundException.java
@@ -1,7 +1,0 @@
-package com.yrlalal.ordermanagementservice.exception;
-
-public class OrderNotFoundException extends RuntimeException {
-    public OrderNotFoundException(String message) {
-        super(message);
-    }
-}
