--- src/main/java/com/yrlalal/ordermanagementservice/v1/entity/OrderItemPK.java
@@ -1,14 +1,0 @@
-package com.yrlalal.ordermanagementservice.v1.entity;
-
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-
-import java.io.Serializable;
-
-public class OrderItemPK implements Serializable {
-    private Integer orderItemId;
-
-    @ManyToOne
-    @JoinColumn(name = "order_id")
-    private Order order;
-}
