--- src/main/java/com/yrlalal/ordermanagementservice/v1/entity/OrderItemStatus.java
@@ -1,5 +1,0 @@
-package com.yrlalal.ordermanagementservice.v1.entity;
-
-public enum OrderItemStatus {
-    NEW, CHECKOUT, ACTIVE, COMPLETED, CANCELLATION_IN_PROGRESS, CANCELLED;
-}
