--- src/main/java/com/yrlalal/ordermanagementservice/v1/entity/OrderStatus.java
@@ -1,5 +1,0 @@
-package com.yrlalal.ordermanagementservice.v1.entity;
-
-public enum OrderStatus {
-    NEW, CHECKOUT, ACTIVE, COMPLETED, CANCELLATION_IN_PROGRESS, CANCELLED;
-}
