--- src/main/java/com/yrlalal/ordermanagementservice/v1/entity/ResponseWrapper.java
@@ -1,2 +1,0 @@
-package com.yrlalal.ordermanagementservice.controlleradvice;public class ResponseWrapper {
-}
