--- src/main/java/com/yrlalal/ordermanagementservice/v1/repository/OrderItemRepository.java
@@ -1,8 +1,0 @@
-package com.yrlalal.ordermanagementservice.v1.repository;
-
-import com.yrlalal.ordermanagementservice.v1.entity.OrderItem;
-import com.yrlalal.ordermanagementservice.v1.entity.OrderItemPK;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OrderItemRepository extends JpaRepository<OrderItem, OrderItemPK> {
-}
