--- src/main/java/com/yrlalal/ordermanagementservice/v1/repository/OrderRepository.java
@@ -1,7 +1,0 @@
-package com.yrlalal.ordermanagementservice.v1.repository;
-
-import com.yrlalal.ordermanagementservice.v1.entity.Order;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OrderRepository extends JpaRepository<Order, String> {
-}
