--- src/main/resources/application.yaml
@@ -1,10 +1,0 @@
-spring:
-  flyway:
-    locations: classpath:db/migration
-    enabled: true
-
-logging:
-  level:
-    org:
-      hibernate:
-        SQL: DEBUG
