--- src/test/java/com/yrlalal/ordermanagementservice/OrderManagementServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.yrlalal.ordermanagementservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class OrderManagementServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
